alexa-airplay/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
python-zeroconf==11.2.0
requests==2.31.0
cryptography==41.0.7
PyYAML==6.0.1
python-dateutil==2.8.2
aiohttp==3.9.1
asyncio-contextmanager==1.0.0
aiofiles==23.2.1
pillow==10.1.0
pydantic==2.5.0
